# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the Voxlab public speaking educational platform",
        "title": "Voxlab API",
        "termsOfService": "http://voxlab.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@voxlab.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/lessons": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a standalone lesson. After creation, link it to a module using\nPOST /api/v1/modules/:id/lessons. Each lesson contains exercises (created separately).\n\n🔒 Requires JWT token (Authorization: Bearer \u003ctoken\u003e)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Create a new lesson",
                "parameters": [
                    {
                        "description": "Lesson data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateLessonRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Lección creada correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.CreateLessonResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos — el título es requerido",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "500": {
                        "description": "Error al crear la lección",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
        "/api/v1/lessons/{id}": {
            "get": {
                "description": "Returns one lesson with its exercises. Use this to load the full lesson content\nincluding all exercise data.\n\n🔓 Public — no authentication required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Get a single lesson by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID (e.g. 1)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lección con ejercicios",
                        "schema": {
                            "$ref": "#/definitions/resources.GetLessonResponse"
                        }
                    },
                    "400": {
                        "description": "ID de lección inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Lección no encontrada",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Changes the title, description, or estimated time of a lesson.\n\n🔒 Requires JWT token (Authorization: Bearer \u003ctoken\u003e)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Update a lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID (e.g. 1)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateLessonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lección actualizada correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.UpdateLessonResponse"
                        }
                    },
                    "400": {
                        "description": "ID de lección inválido o datos incorrectos",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Lección no encontrada",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al actualizar la lección",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently removes a lesson and its exercises.\n⚠️ This action cannot be undone.\n\n🔒 Requires JWT token (Authorization: Bearer \u003ctoken\u003e)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete a lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID (e.g. 1)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lección eliminada correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.DeleteLessonResponse"
                        }
                    },
                    "400": {
                        "description": "ID de lección inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Lección no encontrada",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al eliminar la lección",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/modules/{id}/lessons": {
            "get": {
                "description": "Returns all lessons linked to a specific module, with their exercises.\nLessons are returned through the ModuleLesson pivot and include an order_index.\n\n🔓 Public — no authentication required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "List lessons in a module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Module ID (e.g. 1)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecciones del módulo",
                        "schema": {
                            "$ref": "#/definitions/resources.ListLessonsResponse"
                        }
                    },
                    "400": {
                        "description": "ID de módulo inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Error al obtener las lecciones",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Associates an existing lesson with a module using the ModuleLesson pivot table.\nA lesson can be linked to MULTIPLE modules. This does NOT move or copy the lesson.\n\n🔒 Requires JWT token (Authorization: Bearer \u003ctoken\u003e)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Modules"
                ],
                "summary": "Link a lesson to a module",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Module ID (e.g. 1)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lesson ID to link",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.linkLessonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lección vinculada al módulo correctamente",
                        "schema": {
                            "$ref": "#/definitions/resources.LinkLessonResponse"
                        }
                    },
                    "400": {
                        "description": "ID de módulo inválido o datos incorrectos",
                        "schema": {
                            "$ref": "#/definitions/resources.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Token no proporcionado o inválido",
                        "schema": {
                            "$ref": "#/definitions/resources.UnauthorizedError"
                        }
                    },
                    "404": {
                        "description": "Módulo o lección no encontrados",
                        "schema": {
                            "$ref": "#/definitions/resources.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Error al vincular la lección",
                        "schema": {
                            "$ref": "#/definitions/resources.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreateLessonRequest": {
            "description": "Request body for creating a new lesson",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Aprende a respirar desde el diafragma para proyectar tu voz"
                },
                "estimated_time_seconds": {
                    "type": "integer",
                    "example": 300
                },
                "title": {
                    "type": "string",
                    "example": "Respiración Diafragmática"
                }
            }
        },
...
        "controllers.UpdateLessonRequest": {
            "description": "Request body for updating an existing lesson",
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Técnicas avanzadas de control respiratorio"
                },
                "estimated_time_seconds": {
                    "type": "integer",
                    "example": 600
                },
                "title": {
                    "type": "string",
                    "example": "Respiración Avanzada"
                }
            }
        },
...
        "controllers.linkLessonRequest": {
            "type": "object",
            "properties": {
                "lesson_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.Exercise": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "order_index": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.ExerciseType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ExerciseType": {
            "type": "string",
            "enum": [
                "reading",
                "quiz",
                "audio",
                "oratory_minigame",
                "video",
                "writing"
            ],
            "x-enum-varnames": [
                "ExerciseTypeReading",
                "ExerciseTypeQuiz",
                "ExerciseTypeAudio",
                "ExerciseTypeOratoryMinigame",
                "ExerciseTypeVideo",
                "ExerciseTypeWriting"
            ]
        },
        "models.Lesson": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "estimated_time_seconds": {
                    "type": "integer"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Exercise"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "resources.BadRequestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Solicitud inválida — verifique los campos enviados"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.CreateLessonResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Lesson"
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.DeleteLessonResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.GetLessonResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Lesson"
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error interno del servidor"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "resources.LinkLessonResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.ListLessonsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Lesson"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resources.NotFoundError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Recurso no encontrado"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "resources.UnauthorizedError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Token de autenticación no proporcionado o inválido"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "resources.UpdateLessonResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Lesson"
                },
                "message": {
                    "type": "string",
                    "example": "Operación realizada con éxito"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}